v1.0

+++ SWEN30006_2018S2_PartC/src/mycontroller/DijkstraMinimalPath/Node.java
@@ -42,7 +42,7 @@ public class Node implements Comparable<Node> {
 
 	@Override
 	public int compareTo(Node o) {
-		return Double.compare(this.cost, o.cost);
+		return Integer.compare(this.cost, o.cost);
 	}
 
 }
